2005-02-25  Eric Blake  <ebb9@byu.net>
        * libc/include/time.h (__tzrule_struct): Make offset long, since
        a 16-bit int overflows on a 12-hour offset.
        * libc/sys/linux/include/time.h: Ditto.
        * libc/time/mktime.c (mktime): Use new type of __tzrule.offset.
        * libc/time/mktm_r.c: Ditto.
        * libc/time/gettzinfo.c: Ditto.
        * libc/time/strftime.c (strftime): Fix '%x' to deal with negative
        years.  Fix '%z' to use long, not int.

+++ newlib/libc/include/time.h
@@ -79,7 +79,7 @@ typedef struct __tzrule_struct
   int d;
   int s;
   time_t change;
-  int offset;
+  long offset; /* Match type of _timezone. */
 } __tzrule_type;
 
 typedef struct __tzinfo_struct
